Docky-Talky/requirements.txt
aiohttp==3.9.5
discord.py==2.4.0
fitz==0.0.1.dev2
pdf2image==1.17.0
Pillow==10.4.0
psutil==6.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
Requests==2.32.3
selenium==4.24.0
